feat: add config option for auto-delete-branch-after-merge
